Use newer Learning Resource list cards in Learning Paths lists (#1256)
* Refactor ListDetailsPage ListDetailsComponent to page-component ItemsListing

* Condensed flag for reuse between user list and learning path items

* Remove no items empty message while loading

* Update test path. Variable spacing.

* Margin around ItemsListingComponent
