+++ frontends/mit-open/src/pages/ListDetailsPage/ListDetailsPage.test.ts
@@ -5,7 +5,7 @@ import type {
   PaginatedLearningPathRelationshipList,
 } from "api"
 import { manageListDialogs } from "@/page-components/ManageListDialogs/ManageListDialogs"
-import ItemsListing from "./ItemsListing"
+import ItemsListing from "@/page-components/ItemsListing/ItemsListing"
 import { learningPathsView } from "@/common/urls"
 import {
   screen,
@@ -20,8 +20,10 @@ import { User } from "../../types/settings"
 import { ControlledPromise } from "ol-test-utilities"
 import invariant from "tiny-invariant"
 
-jest.mock("./ItemsListing", () => {
-  const actual = jest.requireActual("./ItemsListing")
+jest.mock("../../page-components/ItemsListing/ItemsListing", () => {
+  const actual = jest.requireActual(
+    "../../page-components/ItemsListing/ItemsListing",
+  )
   return {
     __esModule: true,
     ...actual,
